+++ examples/admin-sdk-app/src/frontend/global.d.ts
@@ -1,3 +1,3 @@
 declare module '*.vue';
 
-import '@shopware-ag/entity-schema-types';
+import '@haokeyingxiao/entity-schema-types';