+++ examples/nuxt-app/app.vue
@@ -26,13 +26,13 @@
 </template>
 
 <script setup lang="ts">
-import '@shopware-ag/meteor-component-library/dist/style.css';
+import '@haokeyingxiao/meteor-component-library/dist/style.css';
 import {
   SwCard,
   SwButton,
   SwBanner,
   SwTextField,
-} from '@shopware-ag/meteor-component-library';
+} from '@haokeyingxiao/meteor-component-library';
 import SwBlockStack from './components/sw-block-stack.vue';
 
 const showBanner = ref();