+++ packages/component-library/src/components/icons-media/mt-icon/mt-icon.vue
@@ -103,7 +103,7 @@ export default defineComponent({
         const iconName = newName.split("-").slice(1).join("-");
 
         import(
-          `./../../../../node_modules/@shopware-ag/meteor-icon-kit/icons/${variant}/${iconName}.svg`
+          `./../../../../node_modules/@haokeyingxiao/meteor-icon-kit/icons/${variant}/${iconName}.svg`
         ).then((iconSvgData) => {
           if (iconSvgData.default) {
             this.iconSvgData = iconSvgData.default;
@@ -126,7 +126,7 @@ export default defineComponent({
 </script>
 
 <style lang="scss">
-@import "node_modules/@shopware-ag/meteor-icon-kit/icons/meteor-icon-kit.scss";
+@import "node_modules/@haokeyingxiao/meteor-icon-kit/icons/meteor-icon-kit.scss";
 
 .mt-icon {
   display: inline-block;