+++ packages/component-library/CHANGELOG.md
@@ -86,7 +86,7 @@
 
 - 7ff2788: Remove unnecessary left border in data table
 - Updated dependencies [8443590]
-  - @shopware-ag/meteor-tokens@0.1.0
+  - @haokeyingxiao/meteor-tokens@0.1.0
 
 ## 3.3.0